requirements.txt
PyMuPDF
cloudinary
psycopg2-binary
requests
python-docx
pdfminer.six
python-dotenv
sentence-transformers
pinecone
flask
flask-cors